requirements.txt
openai==0.28.0
python-dotenv==0.21.1
typing==3.7.4.3
numpy==1.24.2
langchain==0.0.335
streamlit==1.20.0